char32_t not means "unicode codepoint" not an integer
Though this is what we actually mean by UniChar, when we actually want
to output the numeric codepoint it means casts are necessary

+++ tests/test_unicode.cpp
@@ -43,7 +43,8 @@ struct example_unicode
 			{
 				LogError(
 				    "String \"%s\" has unexpected codepoint at index %zu - got 0x%x expected 0x%x",
-				    u8string, i, decoded_codepoints[i], expected_codepoints[i]);
+				    u8string, i, static_cast<uint32_t>(decoded_codepoints[i]),
+				    static_cast<uint32_t>(expected_codepoints[i]));
 				return false;
 			}
 			string2 += decoded_codepoints[i];
